+++ src/main/java/br/com/geradorprova/controller/TagController.java
@@ -1,6 +1,6 @@
 package br.com.geradorprova.controller;
 
-import javax.validation.Valid;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,6 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import br.com.geradorprova.model.Tag;
+import br.com.geradorprova.model.UserSingleton;
+import br.com.geradorprova.model.enumeration.TipoUsuario;
 import br.com.geradorprova.service.TagService;
 
 @Controller
@@ -22,6 +24,12 @@ public class TagController {
 	
 	@GetMapping("/novo")
 	public String tagFormNew(Model model) {
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			
+			return "redirect:/index";
+		}
+		
 		Tag tag = new Tag();
 		model.addAttribute("tag", tag);
 		
@@ -29,7 +37,11 @@ public class TagController {
 	}
 	
 	@PostMapping("/salvar")
-	public String save(@Valid Tag tag, Model model) {
+	public String save(Tag tag, Model model) {
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
 		
 		tagService.save(tag);
 		
@@ -39,6 +51,10 @@ public class TagController {
 	@GetMapping("/deletar/{id}")
 	public String delete(@PathVariable Long id) {
 		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
+		
 		tagService.delete(id);
 		
 		return "redirect:/tag/listar";
@@ -46,6 +62,10 @@ public class TagController {
 	
 	@GetMapping("/editar/{id}")
 	public String formEdit(@PathVariable Long id, Model model) {
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo()) || TipoUsuario.ALUNO.equals(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
 		
 		Tag tag = tagService.findById(id);
 		model.addAttribute("tag", tag);
@@ -55,6 +75,11 @@ public class TagController {
 	
 	@GetMapping("/listar")
 	public String list(Model model) {
+		
+		if(Objects.isNull(UserSingleton.getInstance().getTipo())) {
+			return "redirect:/index";
+		}
+		
 		model.addAttribute("tagList", tagService.list());
 		
 		return "tag/listar.html";
